+++ src/cl_codegen/mod.rs
@@ -1,20 +1,14 @@
-mod printer;
-
 use std::collections::HashMap;
-use crate::codegen as cu_codegen;
-use crate::ast as desc;
-use crate::ast::{BinOpNat, Nat, ScalarTy};
-use crate::c_ast as c;
-use crate::cpp_ast as cpp;
-use crate::c_ast::cpp_to_c_mapper::{CppToCMap, walk_expr, walk_nat, walk_param_decl, walk_ty};
-use crate::c_ast::Stmt::Expr;
-use crate::cpp_ast::{BinOp, Item, ParamDecl, TemplateArg, TemplParam, Ty};
 
-macro_rules! map_list {
-    ($mapper: expr, $method: ident, $list: expr) => {
-        $list.to_vec().into_iter().map(|mut f| $mapper.$method(& f)).collect()
-    };
-}
+use crate::{codegen as cu_codegen, map_list};
+use crate::ast as desc;
+use crate::c_ast as c;
+use crate::c_ast::cpp_to_c_mapper::{CppToCMap, walk_expr};
+use crate::cpp_ast as cpp;
+use crate::cpp_ast::{Item, TemplateArg};
+
+mod printer;
+mod monomorphize_visitor;
 
 pub fn gen_cl (compil_unit: &desc::CompilUnit, idx_checks: bool) -> String {
     let mut cu_program = cu_codegen::gen_items(compil_unit, idx_checks);
@@ -60,8 +54,8 @@ impl<'a> CopyVisitor<'a> {
             });
 
             if let Some(template_fun) = cpp_fun_def {
-                let mangled_function_name = mangle_function_name(fun_identifier.clone(), &template_args);
-                let mut monomorphizer = MonomorphizeVisitor {
+                let mangled_function_name = monomorphize_visitor::mangle_function_name(fun_identifier.clone(), &template_args);
+                let mut monomorphizer = monomorphize_visitor::MonomorphizeVisitor {
                     template_args,
                     template_fun,
                     values_for_names: HashMap::new(),
@@ -87,48 +81,6 @@ impl<'a> CopyVisitor<'a> {
     }
 }
 
-fn mangle_function_name(function_name: String, template_args: &Vec<TemplateArg> ) -> String {
-    let mut mangled_name = function_name.clone();
-    template_args.iter().map(|f| {
-        match f {
-            TemplateArg::Expr(expr) => {
-                if let cpp::Expr::Nat(nat) = expr {
-                    match nat {
-                        Nat::Lit(size) => {size.to_string()},
-                        _ => { panic!("Trying to Mangle Nat which is not Lit") }
-                    }
-                } else {
-                    panic!("Trying to mangle Exprssion which is not a nat")
-                }
-            }
-            TemplateArg::Ty(ty) => {
-                match ty {
-                    Ty::Scalar(scalar_ty) => {scalar_ty_to_name(scalar_ty)}
-                    Ty::Atomic(scalar_ty) => {scalar_ty_to_name(scalar_ty)}
-                    _ => {panic!("Unmapped Type!")}
-                }
-            }
-        }
-    }).for_each(|str| mangled_name.push_str(&*str));
-
-    mangled_name
-}
-
-fn scalar_ty_to_name(ty: &cpp::ScalarTy) -> String{
-    match ty {
-        cpp::ScalarTy::Void => {"void"}
-        cpp::ScalarTy::I32 => {"i32"}
-        cpp::ScalarTy::U32 => {"u32"}
-        cpp::ScalarTy::F32 => {"f32"}
-        cpp::ScalarTy::F64 => {"f64"}
-        cpp::ScalarTy::Bool => {"bool"}
-        cpp::ScalarTy::Gpu => {"gpu"}
-        cpp::ScalarTy::SizeT => {"size-t"}
-        cpp::ScalarTy::Memory => {"memory"}
-        _ => {panic!("Unmapped Type!")}
-    }.to_string()
-}
-
 impl<'a> CppToCMap for CopyVisitor<'a> {
     fn map_expr(&mut self, expr: &cpp::Expr) -> c::Expr {
         match expr {
@@ -138,95 +90,8 @@ impl<'a> CppToCMap for CopyVisitor<'a> {
                 } else {
                     walk_expr(self, expr)
                 }
-            },
+            }
             _ => walk_expr(self, expr)
         }
     }
-}
-
-struct MonomorphizeVisitor<'b> {
-    template_args: Vec<TemplateArg>,
-    template_fun: &'b Item,
-    values_for_names: HashMap<String, TemplateArg>
-}
-
-impl<'b> MonomorphizeVisitor<'b> {
-    fn find_names_for_template_args(&mut self) {
-        if let Item::FunDef { templ_params, .. } = self.template_fun {
-            for i in 0 .. templ_params.len() {
-                if let Some(param) = templ_params.get(i) {
-                    let param_name = match param {
-                        TemplParam::Value { param_name, ty } => {param_name.clone()}
-                        TemplParam::TyName { name } => {name.clone()}
-                    };
-                    if let Some(template_args) = self.template_args.get(i) {
-                        self.values_for_names.insert(param_name, template_args.clone());
-                    }
-                }
-            }
-        } else {
-            panic!("No Monomorph for include!");
-        }
-    }
-}
-
-impl<'b> CppToCMap for MonomorphizeVisitor<'b> {
-    fn map_item(&mut self, item: &Item) -> Option<c::Item> {
-        self.find_names_for_template_args();
-        match item {
-            cpp::Item::FunDef {
-                name,
-                params,
-                templ_params,
-                ret_ty,
-                body,
-                is_dev_fun } => {
-                // Templated Functions are mapped only when their apply is mapped. Therefor we filter them here
-                if (!templ_params.is_empty()) {
-                    Some(c::Item::FunDef {
-                        name: mangle_function_name(name.clone(), &self.template_args),
-                        params: map_list!(self, map_param_decl, params),
-                        ret_ty: self.map_ty(ret_ty),
-                        body: self.map_stmt(body),
-                        is_dev_fun: is_dev_fun.clone(),
-                    })
-                } else {
-                    panic!("I think you're using this the wrong way");
-                }
-            }
-            _ => {
-                panic!("I think you're using this the wrong way");
-            }
-        }
-    }
-
-    fn map_nat(&mut self, nat: &Nat) -> Nat {
-        match nat {
-            Nat::Ident(ident) => {
-                if let Some(template_param) = self.values_for_names.get(&*ident.name) {
-                    match template_param {
-                        TemplateArg::Expr(ex) => {
-                            if let cpp::Expr::Nat(template_nat) = ex {
-                                template_nat.clone()
-                            }
-                            else {
-                                panic!("Trying to map non Nat Template Param to Nat");
-                            }
-                        }
-                        TemplateArg::Ty(_) => {panic!("Template Param missmatch trying to assing type to nat!")}
-                    }
-                } else {
-                    walk_nat(self,nat)
-                }
-            }
-            Nat::BinOp(op, lhs, rhs) => {
-                Nat::BinOp(
-                    op.clone(),
-                    Box::new(self.map_nat(lhs)),
-                    Box::new(self.map_nat(rhs)))
-            }
-            _ => walk_nat(self, nat)
-        }
-    }
-
 }
